+++ EssentialFeed/Feed Presentation/FeedPresenter.swift
@@ -1,0 +1,83 @@
+//
+//  FeedPresenter.swift
+//  EssentialFeed
+//
+//  Created by Christos Tzimas on 9/7/25.
+//
+
+import Foundation
+
+public protocol FeedView {
+    func display(_ viewModel: FeedViewModel)
+}
+
+public struct FeedViewModel {
+    public let feed: [FeedImage]
+}
+
+public protocol FeedLoadingView {
+    func display(_ viewModel: FeedLoadingViewModel)
+}
+
+public struct FeedLoadingViewModel {
+    public let isLoading: Bool
+}
+
+public protocol FeedErrorView {
+    func display(_ viewModel: FeedErrorViewModel)
+}
+
+public struct FeedErrorViewModel {
+    
+    public let message: String?
+    
+    public static var noError: FeedErrorViewModel {
+        FeedErrorViewModel(message: nil)
+    }
+    
+    public static func error(message: String) -> FeedErrorViewModel {
+        FeedErrorViewModel(message: message)
+    }
+}
+
+public final class FeedPresenter {
+     
+    private let feedView: FeedView
+    private let loadingView: FeedLoadingView
+    private let errorView: FeedErrorView
+
+    public init(feedView: FeedView, loadingView: FeedLoadingView, errorView: FeedErrorView) {
+        self.feedView = feedView
+        self.loadingView = loadingView
+        self.errorView = errorView
+    }
+    
+    public static var title: String {
+        NSLocalizedString("FEED_VIEW_TITLE",
+                          tableName: "Feed",
+                          bundle: Bundle(for: FeedPresenter.self),
+                          comment: "Title for the feed view")
+    }
+    
+    private var feedLoadError: String {
+        NSLocalizedString("FEED_VIEW_CONNECTION_ERROR",
+                          tableName: "Feed",
+                          bundle: Bundle(for: FeedPresenter.self),
+                          comment: "Error message when loading fails")
+    }
+    
+    public func didStartLoadingFeed() {
+        errorView.display(.noError)
+        loadingView.display(FeedLoadingViewModel(isLoading: true))
+    }
+    
+    public func didFinishLoadingFeed(with feed: [FeedImage]) {
+        feedView.display(FeedViewModel(feed: feed))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+    
+    public func didFinishLoadingFeed(with error: Error) {
+        errorView.display(.error(message: feedLoadError))
+        loadingView.display(FeedLoadingViewModel(isLoading: false))
+    }
+}
