+++ EssentialFeed/Feed Presentation/pt-BR.lproj/Feed.strings
